# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Yzu视频 Premium Video Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subscriptions:
    get:
      operationId: listSubscriptions
      tags: [subscriptions]
      summary: List available subscription plans
      responses:
        "200":
          description: Subscription plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
    post:
      operationId: createSubscription
      tags: [subscriptions]
      summary: Create subscription plan (owner)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
  /subscriptions/{id}:
    patch:
      operationId: updateSubscription
      tags: [subscriptions]
      summary: Update subscription plan (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubscriptionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
    delete:
      operationId: deleteSubscription
      tags: [subscriptions]
      summary: Delete subscription plan (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /subscriptions/{id}/purchase:
    post:
      operationId: purchaseSubscription
      tags: [subscriptions]
      summary: Purchase subscription using wallet balance
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSubscription"
...
  /topups:
    get:
      operationId: listMyTopups
      tags: [topups]
      summary: List current user topup history
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Topup history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupList"
    post:
      operationId: createTopup
      tags: [topups]
      summary: Create a topup request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopupInput"
      responses:
        "201":
          description: Topup created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topup"
  /topups/all:
    get:
      operationId: listAllTopups
      tags: [topups]
      summary: List all topups (owner)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, denied]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: All topups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopupList"
  /topups/{id}/confirm:
    patch:
      operationId: confirmTopup
      tags: [topups]
      summary: Confirm a topup (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topup"
  /topups/{id}/deny:
    patch:
      operationId: denyTopup
      tags: [topups]
      summary: Deny a topup (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Topup"
  /topups/{id}:
    delete:
      operationId: deleteTopup
      tags: [topups]
      summary: Delete a topup record (owner)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

  # ── TRANSACTIONS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, isBanned, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin, owner]
        avatar:
          type: ["string", "null"]
        isBanned:
          type: boolean
        walletBalance:
          type: number
        totalTopup:
          type: number
        totalSpent:
          type: number
        createdAt:
          type: string
          format: date-time
        activeSubscription:
          $ref: "#/components/schemas/UserSubscription"
...
    Subscription:
      type: object
      required: [id, name, price, durationDays, isActive]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
    SubscriptionInput:
      type: object
      required: [name, price, durationDays]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
    SubscriptionUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
    UserSubscription:
      type: object
      required: [id, userId, subscriptionId, startDate, endDate, isActive]
      properties:
        id:
          type: string
        userId:
          type: string
        subscriptionId:
          type: string
        subscription:
          $ref: "#/components/schemas/Subscription"
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        isActive:
          type: boolean

    # ── TOPUP ───────────────────────────────────────────────────────────────
    Topup:
      type: object
      required: [id, userId, amount, status, createdAt]
      properties:
        id:
          type: string
        userId:
          type: string
        user:
          $ref: "#/components/schemas/User"
        amount:
          type: number
        paymentProof:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, denied]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TopupInput:
      type: object
      required: [amount, paymentProof]
      properties:
        amount:
          type: number
        paymentProof:
          type: string
    TopupList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Topup"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── TRANSACTION ─────────────────────────────────────────────────────────
...