# lib/api-zod/src/generated/types/subscription.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yzu视频 Premium Video Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Subscription {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  durationDays: number;
  isActive: boolean;
}