# lib/api-zod/src/generated/types/topup.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yzu视频 Premium Video Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TopupStatus } from './topupStatus';
import type { User } from './user';

export interface Topup {
  id: string;
  userId: string;
  user?: User;
  amount: number;
  /** @nullable */
  paymentProof?: string | null;
  status: TopupStatus;
  createdAt: Date;
  updatedAt?: Date;
}