# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yzu视频 Premium Video Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserSubscription } from './userSubscription';

export interface User {
  id: string;
  username: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatar?: string | null;
  isBanned: boolean;
  walletBalance?: number;
  totalTopup?: number;
  totalSpent?: number;
  createdAt: Date;
  activeSubscription?: UserSubscription;
}