# lib/api-zod/src/generated/types/userSubscription.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Yzu视频 Premium Video Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Subscription } from './subscription';

export interface UserSubscription {
  id: string;
  userId: string;
  subscriptionId: string;
  subscription?: Subscription;
  startDate: Date;
  endDate: Date;
  isActive: boolean;
}